other: Seperated CSS from index.js into new stylesheet
