+++ app/index.js
@@ -1,14 +1,6 @@
 import { StatusBar } from "expo-status-bar";
-import { StyleSheet, Text, View } from "react-native";
-
-const styles = StyleSheet.create({
-    "container": {
-        "flex": 1,
-        "backgroundColor": "#fff",
-        "alignItems": "center",
-        "justifyContent": "center",
-    },
-});
+import { Text, View } from "react-native";
+import { styles } from "./style.js";
 
 export default function App() {
     return (
