+++ app/style.js
@@ -1,0 +1,12 @@
+import { StyleSheet } from "react-native";
+
+const styles = StyleSheet.create({
+    "container": {
+        "flex": 1,
+        "backgroundColor": "#fff",
+        "alignItems": "center",
+        "justifyContent": "center",
+    },
+});
+
+export { styles };
